+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 1.1.0
-Add option to change decimal seperators on all locales.
+Add the `patchNumberSeperators` function to change decimal and group seperators on all locales.
 
 ## 1.0.1