+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.0
+Add the `patchNumberSeperators` function to change decimal and group seperators on all locales.
+
 ## 1.0.1
 
 Authors updated.